updated libft

+++ libft/ft_memmove.c
@@ -1,6 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memmove.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ibenaiss <ibenaiss@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/04 18:42:30 by ibenaiss          #+#    #+#             */
+/*   Updated: 2023/11/04 18:49:22 by ibenaiss         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-void *ft_memmove(void *dest, const void *src, int n)
+void *ft_memmove(void *dest, const void *src, size_t n)
 {
     unsigned char *d;
     const unsigned char *s;
@@ -24,5 +36,4 @@ void *ft_memmove(void *dest, const void *src, int n)
         }
     }
     return (dest);
-
 }
