+++ libft/ft_strnstr.c
@@ -1,4 +1,18 @@
-char *ft_strnstr(const char *big, const char *little, size_t len)
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strnstr.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ibenaiss <ibenaiss@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/04 18:45:41 by ibenaiss          #+#    #+#             */
+/*   Updated: 2023/11/04 18:45:42 by ibenaiss         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "libft.h"
+
+char *ft_strnstr(const char *big, const char *little, int len)
 {
     const char *ptr;
 
@@ -15,5 +29,3 @@ char *ft_strnstr(const char *big, const char *little, size_t len)
     }
     return NULL;
 }
-
-
